jpress3

+++ module-article/module-article-service-api/src/main/java/io/jpress/module/article/service/ArticleService.java
@@ -106,7 +106,7 @@ public interface ArticleService extends JbootServiceJoiner {
 
     public boolean doChangeStatus(long id, String status);
 
-    public int findCountByStatus(String status);
+    public Long findCountByStatus(String status);
 
     public Article findFirstBySlug(String slug);
